refactor(brain): let the dream pass read as the stages it runs
`dream.ts` was 1052 lines of orchestration fused with the mechanics of
every stage it orchestrates (cognitive complexity 391, MI -61.0). The
pass genuinely has stages - scan, plan topics, plan retires, apply,
report - and each one has its own inputs and its own decision, so each
one now has its own module named for what it decides. The coordinator
drops to 217 lines and cognitive complexity 18.

The split separates deciding from writing: `dream-apply.ts` owns the
ORDER of writes that keeps the on-disk invariants true and decides
nothing, while the two planning modules decide and write nothing.

`dream.ts` keeps its path and its exact prior surface; no importer
changed and no test needed an import-path edit.

One coverage regression appeared and was closed rather than excluded:
moving the rename out of `dream.ts` made the new apply module
write-capable, and the vault-guard census correctly failed on it. The
fix was to make the guard honest - the apply step now asserts vault
write identity on entry - not to add an exclusion. It is a runtime
no-op, since the workrun open pins that same root immediately before.

Deleted while moving: a planning function that was a pure stub, an
unused parameter, and a `void basename` whose comment claimed a barrel
re-export that never existed. Consolidated: a preference-id prefix strip
open-coded four times in three different spellings, two near-identical
signal-scan loops, and a retain-pinned block written three times.

One latent trap avoided: hoisting the log path into the summary would
have called the vault-write assertion on the dry-run path, where it had
never run before. The summary takes a caller-built discriminant instead,
so a dry run still touches nothing.
